# Perf: Improve perf of SqlDateTimeToDateTime