@Wraith2 2021-02-16
Nice catch. Didn't see it among the other CI failures.
Reworked slightly to deal with the bounds and I've added a uint range check trick that wasn't worth using before (regressed it by 4%) but now does seem worthwhile. The range checks work by knowing that negative integers cast to extremely large positive uints so if we do zero based checks we only have to check 0 to max and not a min check.

|   Method |      Mean |    Error |   StdDev | Ratio |
|--------- |----------:|---------:|---------:|------:|
| Original | 258.05 ms | 2.267 ms | 2.010 ms |  1.00 |
|      New |  85.68 ms | 0.589 ms | 0.551 ms |  0.33 |